【蜜兔】  体统扣减商户余额改sql处理优化

+++ yshop-shop/src/main/java/co/yixiang/modules/shop/service/mapper/UserSysMapper.java
@@ -41,7 +41,7 @@ public interface UserSysMapper extends CoreMapper<User> {
      * @param extractPrice
      */
     @Update("update `user` set withdrawal_amount = withdrawal_amount - #{extractPrice} where id = #{id}")
-    void updateWithdrawalAmountSub(@Param("id") Long id, @Param("extractPrice") BigDecimal extractPrice);
+    void updateWithdrawalAmountSub(@Param("id") Integer id, @Param("extractPrice") BigDecimal extractPrice);
 
     /**
      * 更新商户金额
